Don't allow creation, edit and delete for unauthorized users

+++ project.py
@@ -205,7 +205,12 @@ def restaurantsJSON():
 @app.route('/restaurant/')
 def showRestaurants():
   restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-  return render_template('restaurants.html', restaurants = restaurants)
+
+  if 'username' in login_session:
+    return render_template('restaurants.html', restaurants = restaurants)
+  else:
+    return render_template('publicrestaurants.html', restaurants = restaurants)
+
 
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
@@ -227,6 +232,10 @@ def editRestaurant(restaurant_id):
   if 'username' not in login_session:
       return redirect(url_for('showLogin'))
   editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+  if editedRestaurant.user_id != login_session['user_id']:
+    return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. You may only edit a restaurant you created.');}</script><body onload='myFunction()''>"
+
   if request.method == 'POST':
       if request.form['name']:
         editedRestaurant.name = request.form['name']
@@ -242,6 +251,10 @@ def deleteRestaurant(restaurant_id):
   if 'username' not in login_session:
       return redirect(url_for('showLogin'))
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+  if restaurantToDelete.user_id != login_session['user_id']:
+    return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. You may only delete a restaurant you created.');}</script><body onload='myFunction()''>"
+
   if request.method == 'POST':
     session.delete(restaurantToDelete)
     flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -256,8 +269,13 @@ def deleteRestaurant(restaurant_id):
 def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-    return render_template('menu.html', items = items, restaurant = restaurant)
 
+    if 'username' in login_session:
+        if login_session['user_id'] == restaurant.user_id:
+            return render_template('menu.html', items = items, restaurant = restaurant)
+
+    creator = getUserInfo(restaurant.user_id)
+    return render_template('publicmenu.html', items = items, restaurant = restaurant, creator = creator)
 
 
 #Create a new menu item
@@ -266,6 +284,10 @@ def newMenuItem(restaurant_id):
   if 'username' not in login_session:
       return redirect(url_for('showLogin'))
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+  if restaurant.user_id != login_session['user_id']:
+      return "<script>function myFunction() {alert('You are not authorized to add menu items from this restaurant. You may only add to a restaurant you created.');}</script><body onload='myFunction()''>"
+
   if request.method == 'POST':
       newItem = MenuItem(name = request.form['name'],
                          description = request.form['description'],
@@ -287,6 +309,10 @@ def editMenuItem(restaurant_id, menu_id):
         return redirect(url_for('showLogin'))
     editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+    if restaurant.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to edit menu items from this restaurant. You may only edit from a restaurant you created.');}</script><body onload='myFunction()''>"
+
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -311,6 +337,10 @@ def deleteMenuItem(restaurant_id,menu_id):
         return redirect(url_for('showLogin'))
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one()
+
+    if restaurant.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to delete menu items from this restaurant. You may only delete from a restaurant you created.');}</script><body onload='myFunction()''>"
+
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
